model2/requirements.txt
python3
torch
transformers
Pillow
matplotlib
numpy
requests
